add a groups reaper to remove non existing groups
burrow currently keeps reporting lag for non existing consumers.

The only way to remove groups from burrow automatically is configuring
expire-group, which is not ideal as it can conflict with consumer with
no members.

This PR introduces a go routine to get the existing consumer groups from
Kafka, and compare it against burrow consumers to reap the non existing ones.