Fix off-by-one error in backfill consumer
Much related to linkedin/Burrow#616 but for the actual
`__consumer_offsets` message!

Currently, all backfilled partitions are skipping the final message.

This is not an issue if the consumer group is busy and committing often,
however, if that is not the case it can lead to `Burrow` reporting (a
single consumer group topic partition) as being with status `STOP`
erroneously. This behaviour persists across restarts if no new messages
have been produced on the given `__consumer_offsets` partition.

Upstreamed in: linkedin/Burrow#677

Signed-off-by: Christian Gregg <christian@bissy.io>